+++ src/ui/toolbar/mobile/ViewSection.tsx
@@ -4,6 +4,7 @@ import { enterVr } from '../../../scene/xr/xrStore'
 import { captureThumb } from '../../../state/storage/slotThumbs'
 import { useStore } from '../../../state/store'
 import { PresentationSetup } from '../../presentation/PresentationSetup'
+import { DayNightClipSetup } from '../../scene/DayNightClipSetup'
 import { Icon } from '../icons'
 import { Item, Section } from './parts'
 
@@ -35,6 +36,7 @@ export function ViewSection({
   const fPanoTour = useFeature('panoTour')
   const fTwoPointPerspective = useFeature('twoPointPerspective')
   const fParallelProjection = useFeature('parallelProjection')
+  const fDayNightClip = useFeature('dayNightClip')
 
   return (
     <Section id="view" title="View" icon="Orbit" activeId={activeId}>
@@ -164,6 +166,7 @@ export function ViewSection({
               onClick={act(() => s.getState().setTouring('views'))}
             />
           ) : null}
+          {fSavedViews && savedViews.length > 1 && fDayNightClip ? <DayNightClipSetup /> : null}
         </>
       ) : null}
       {!roomEditorActive &&